docker/ticker_fetcher/requirements.txt
tgedr-nihao
